Initial commit of a survival power analysis SAS macro.
This SAS macro runs a power analysis for the median progression free survival (PFS) using proc power and the log rank test. Inputs needed are the median PFS for the control group and a string to the median PFS on wants to assess for the treatment group, the power (default 0.8), alpha level (default 0.05), number of sides (default 2), accrual time, and followup time.